[Refactor] Notice Authorization Add #80
